requirements.txt
flask
flask_cors
python-binance
requests
numpy